Rename marketo.rb to marketo/lead.rb
I'll implement input plugin for marketo activity log.
